Renamed libifconfig to libifc.
