+++ Makefile.base
@@ -1,16 +1,16 @@
 # $FreeBSD$
 
 PACKAGE=	lib${LIB}
-LIB=		ifconfig
+LIB=		ifc
 
 SHLIBDIR?=	/lib
 SHLIB_MAJOR=	1
-SRCS=		libifconfig.c libifconfig_internal.c
+SRCS=		libifc.c libifc_internal.c
 
 INCSDIR=	${INCLUDEDIR}
-INCS=		libifconfig.h
+INCS=		libifc.h
 
-#MAN=		libifconfig.3
+#MAN=		libifco.3
 
 CFLAGS+= -I${.CURDIR}
